# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Portal de ejemplo para utilizar las apis del Ecommerce para buscar productos",
    "version": "1.0.7",
    "title": "Swagger Ecommerce Search",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    {
      "name": "Category",
      "description": "List of all categories"
    },
    {
      "name": "Brand",
      "description": "List of all brands"
    },
    {
      "name": "Product",
      "description": "All about the product list"
    }
  ],
  "schemes": ["http", "https"],
  "paths": {
    "/category": {
      "get": {
        "tags": ["Category"],
        "summary": "List of all categories",
        "description": "Returns a list of categories",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      }
    },
    "/brand": {
      "get": {
        "tags": ["Brand"],
        "summary": "List of all brands",
        "description": "Returns a list of brands",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Brand"
              }
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "tags": ["Product"],
        "summary": "List of all products",
        "description": "Returns a list of products",
        "produces": ["application/json"],
        "parameters": [
            {
                "name": "nombre",
                "in": "query",
                "description": "Name of Product",
                "required": false,
                "type": "string"
            },
            {
                "name": "precio_min",
                "in": "query",
                "description": "Prices greater than or equal to this",
                "type": "number",
                "required": false,
                "format": "double"
            },
            {
                "name": "precio_max",
                "in": "query",
                "description": "Prices lower than this",
                "type": "number",
                "required": false,
                "format": "double"
            },
            {
                "name": "categoria_id",
                "in": "query",
                "description": "Category Id to search",
                "required": false,
                "type": "integer"
            },
            {
                "name": "marca_id",
                "in": "query",
                "description": "Brand Id to search",
                "required": false,
                "type": "integer"
            },
            {
                "name": "en_oferta",
                "in": "query",
                "description": "Show only products on sale. Send 'true' or nothing",
                "required": false,
                "type": "string"
            },
            {
                "name": "popularidad",
                "in": "query",
                "description": "Sort by popularity from highest to lowest. Send 'true' or nothing",
                "required": false,
                "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      }
    },
    "/product/searchChatGPT":{
      "get": {
        "tags": ["Product"],
        "summary": "List of all products",
        "description": "Returns a list of products",
        "produces": ["application/json"],
        "parameters": [
            {
                "name": "consulta",
                "in": "query",
                "description": "Consulta para ChatGPT",
                "required": true,
                "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      }
    },
    "/product/{productId}": {
      "get": {
        "tags": ["Product"],
        "summary": "Find product by ID",
        "description": "For valid response try integer IDs with value >= 1. Other values will generated exceptions",
        "operationId": "getProductById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "nombre": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Brand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "nombre": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Brand"
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "descripcion": {
          "type": "string"
        },
        "descuento": {
          "type": "string"
        },
        "en_oferta": {
            "type": "integer",
            "format": "int64"
        },
        "nombre": {
          "type": "string"
        },
        "popularidad": {
          "type": "string"
        },
        "precio": {
          "type": "string"
        },
        "stock": {
            "type": "integer",
            "format": "int64"
        }
      },
      "xml": {
        "name": "Product"
      }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}